chore: rename main index file to index.ts
