+++ packages/app-banner/src/index.ts
@@ -1,0 +1,1 @@
+export { default } from './app-banner'
