+++ packages/content-sharing/src/content-sharing.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react'
 
-import ContentSharing from '.'
+import ContentSharing from './content-sharing'
 
 export default {
   title: 'content-sharing / ContentSharing',
