+++ packages/content-sharing/src/index.ts
@@ -1,0 +1,2 @@
+export * from './content-sharing'
+export { default } from './content-sharing'
