+++ packages/directions-finder/src/index.ts
@@ -1,0 +1,1 @@
+export { default } from './direction-buttons'
