--- packages/directions-finder/src/index.tsx
@@ -1,3 +1,0 @@
-import DirectionButtons from './direction-buttons'
-
-export default DirectionButtons
