+++ packages/drawer-button/src/index.ts
@@ -1,0 +1,2 @@
+export * from './drawer-button'
+export { default } from './drawer-button'
