+++ packages/intersection-observer/src/index.ts
