+++ packages/pricing/src/index.ts
@@ -1,0 +1,4 @@
+export * from './fixed-pricing-v2'
+export { default as FixedPricingV2 } from './fixed-pricing-v2'
+export * from './pricing'
+export { default } from './pricing'
