+++ packages/recommended-contents/src/index.ts
@@ -1,0 +1,1 @@
+export { default } from './recommended-contents'
