+++ packages/recommended-contents/src/recommended-contents.stories.tsx
@@ -1,8 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react'
 
 import { contents } from './mocks/recommended-contents.sample.json'
-
-import RecommendedContents from '.'
+import RecommendedContents from './recommended-contents'
 
 export default {
   title: 'recommended-contents / RecommendedContents',
