+++ packages/search/src/search.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react'
 
-import Search from '.'
+import Search from './search'
 
 export default {
   title: 'search / Search',
