+++ packages/static-map/src/index.ts
@@ -1,0 +1,2 @@
+export * from './static-map'
+export { default } from './static-map'
