+++ packages/static-map/src/static-map.tsx
