+++ packages/static-page-contents/src/index.ts
@@ -1,0 +1,1 @@
+export * from './static-page-contents'
