+++ packages/static-page-contents/src/static-page-contents.tsx
