+++ packages/triple-fallback-action/src/index.ts
@@ -1,0 +1,1 @@
+export * from './triple-fallback-action'
