+++ packages/triple-media/src/index.ts
@@ -1,0 +1,2 @@
+export * from './triple-media'
+export { default } from './triple-media'
